+++ src/test/java/com/codeborne/selenide/impl/CollectionElementTest.java
@@ -9,15 +9,13 @@ import org.assertj.core.api.WithAssertions;
 import org.junit.jupiter.api.Test;
 import org.openqa.selenium.WebElement;
 
-import java.util.List;
-
-import static java.util.Arrays.asList;
+import static com.codeborne.selenide.Mocks.mockCollection;
 import static java.util.Collections.singletonList;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 class CollectionElementTest implements WithAssertions {
-  private Driver driver = new DriverStub();
+  private final Driver driver = new DriverStub();
 
   @Test
   void wrap() {
@@ -34,15 +32,12 @@ class CollectionElementTest implements WithAssertions {
 
   @Test
   void getWebElement() {
-    WebElementsCollection collection = mock(WebElementsCollection.class);
     WebElement mockedWebElement1 = mock(WebElement.class);
     WebElement mockedWebElement2 = mock(WebElement.class);
-    List<WebElement> listOfMockedElements = asList(mockedWebElement1, mockedWebElement2);
-    when(collection.getElements()).thenReturn(listOfMockedElements);
+    WebElementsCollection collection = mockCollection("", mockedWebElement1, mockedWebElement2);
     CollectionElement collectionElement = new CollectionElement(collection, 1);
 
-    assertThat(collectionElement.getWebElement())
-      .isEqualTo(mockedWebElement2);
+    assertThat(collectionElement.getWebElement()).isEqualTo(mockedWebElement2);
   }
 
   @Test